+++ CHANGELOG.md
@@ -12,13 +12,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `0.1.x` : GitHub Actions
 - `0.2.x` : Simple Tic Tac Toe (3 x 3)
 - `0.3.x` : Implement ultimate edition (9 x 9)
+- `0.4.x` : Add single player mode
+- `0.5.x` : Implement push notifications
+- `0.6.x` : UI Revamp
+- `0.7.x` : Misc Improvements
+- `1.0.x` : Initial release after user-tests
 
 <br>
 
 # Releases
 <!-- @LatestFirst -->
 
-## [0.3.2] - 11/11/23
+## [0.4.0] - ???
+- UMA-25: Add single player option to log in
+- UMA-27: In standard mode, play computer's move after user has placed
+- UMA-28: In ultimate mode, play computer's move after user has placed
+
+## [0.3.2] - 15/11/23
 - UMA-18: Add replay button
 
 ## [0.3.1] - 11/11/23
@@ -64,3 +74,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.3.0]: https://github.com/jrsmth/ultima/compare/0.2.1...0.3.0
 [0.3.1]: https://github.com/jrsmth/ultima/compare/0.3.0...0.3.1
 [0.3.2]: https://github.com/jrsmth/ultima/compare/0.3.1...0.3.2
+[0.4.0]: https://github.com/jrsmth/ultima/compare/0.3.2...0.4.0